Car Selection
Added a car selection menu.  Now when you select a car and a track it calls the track scene and exports the selected car scene as a global variable. This means each track will need a spawn point and check that global var and spawn it. Works pretty good tho imo.